requirements.txt
openpyxl==3.1.2
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
peewee==3.16.3
loguru==0.7.0
python-dotenv==1.0.0